removed ConfigModelBean  code migrated to ConfigServiceBean  usage changed to Service at apropriate places
